Add task_helper.rb
