+++ config/locales/ja.yml
@@ -37,6 +37,7 @@ ja:
     not_yet_arrived: 未着手
     underway: 着手中
     complete: 完了
+    search: 検索
