+++ config/routes.rb
@@ -1,5 +1,4 @@
 Rails.application.routes.draw do
   root 'tasks#index'
   resources :tasks
-
 end
