+++ db/migrate/20190916211357_add_status_to_tasks.rb
@@ -1,0 +1,5 @@
+class AddStatusToTasks < ActiveRecord::Migration[6.0]
+  def change
+    add_column :tasks, :status, :string, null: false, default: "not_yet_arrived"
+  end
+end
